# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Retriva Core API
  description: |
    Retriva Core provides APIs for document ingestion, chunking, indexing,
    retrieval, and OpenAI-compatible RAG chat completions.

    The Core runs as two processes:
    * **Ingestion API** (port 8000): Document processing pipeline, metadata, jobs, artifacts.
    * **OpenAI Chat API** (port 8001): OpenAI-compatible `/v1/chat/completions` for RAG.
  version: 1.5.1
  contact:
    name: Andrea Marson
    email: am.dev.75@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8000
    description: Retriva Core — Ingestion API
  - url: http://localhost:8001
    description: Retriva Core — OpenAI Chat API
...
paths:
...
  /api/v2/jobs/{job_id}:
    get:
      tags: [Jobs v2]
      summary: Get v2 job status
      operationId: getJobV2
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details with pipeline stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponseV2"

  # --- Metadata API (v2) ---
...
  /api/v2/crm/archive/jobs:
    get:
      tags:
      - crm-archive
      summary: List Archived Jobs
      description: List archived jobs (tenant-scoped; filter/sort/paginate).
      operationId: list_archived_jobs_api_v2_crm_archive_jobs_get
      parameters:
      - name: session_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      - name: state
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: order
        in: query
        required: false
        schema:
          type: string
          default: desc
          title: Order
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/crm/archive/jobs/{job_id}:
    get:
      tags:
      - crm-archive
      summary: Get Archived Job
      description: 'Durable record of one terminal job (audit metadata + bindings +

        archive status).'
      operationId: get_archived_job_api_v2_crm_archive_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v2/crm/jobs/{job_id}:
    get:
      tags:
      - crm-assistant
      summary: Get Job Status
      description: 'Get the status of a qualification job.


        Milestone A: when the live in-memory job record has disappeared

        (restart, container recreation) but a durable archive exists, the

        ARCHIVED terminal result is returned — never a synthetic LOST

        merely because process memory was cleared.'
      operationId: get_job_status_api_v2_crm_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/crm/jobs/{job_id}/cancel:
    post:
      tags:
      - crm-assistant
      summary: Cancel Job
      description: 'Request cancellation of a qualification job.


        When ``session_id`` is supplied (the chat path always supplies the

        trusted session), the job must belong to that session — a foreign

        session''s cancel request is rejected with 403.  Without the parameter

        (operational/direct use), cancellation proceeds as before.'
      operationId: cancel_job_api_v2_crm_jobs__job_id__cancel_post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: session_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v2/crm/sessions/{session_id}/jobs:
    get:
      tags:
      - crm-assistant
      summary: List Session Jobs
      description: 'List qualification jobs for a chat session (session-scoped).


        Enables cross-turn job resumption: a later chat turn can discover the

        jobs of ITS OWN session server-side, without relying on the LLM

        remembering job IDs from conversation text.  Only jobs of the requested

        session are returned — other sessions'' jobs are never exposed.


        Milestone A: archived terminal jobs of the session that are no

        longer in the in-memory registry are merged (``archived: true``)

        so resumption still finds them after restarts/recreation.'
      operationId: list_session_jobs_api_v2_crm_sessions__session_id__jobs_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    JobResponseV2:
      type: object
      properties:
        job_id: { type: string }
        status: { type: string }
        source: { type: string }
        job_type: { type: string }
        current_stage: { type: string }
        stages_completed: { type: array, items: { type: string } }
        stage_detail: { type: string, description: "Human-readable detail about the current stage (e.g. 'parsing chunk 7/14')" }
        progress: { type: number, format: float, description: "Optional 0.0–1.0 progress indicator for the current stage" }
        created_at: { type: string }
        updated_at: { type: string }
        error: { type: string }
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JobStatusResponse:
      properties:
        job_id:
          type: string
          title: Job Id
        state:
          type: string
          title: State
        progress:
          type: number
          title: Progress
          default: 0.0
        stage_detail:
          type: string
          title: Stage Detail
          default: ''
        candidate_count:
          type: integer
          title: Candidate Count
          default: 0
        result_count:
          type: integer
          title: Result Count
          default: 0
        artifact_ids:
          items:
            type: string
          type: array
          title: Artifact Ids
        icp_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Icp Id
        icp_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Icp Version
        portfolio_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Portfolio Id
        portfolio_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Portfolio Version
        analysis_mode:
          anyOf:
          - type: string
          - type: 'null'
          title: Analysis Mode
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        warnings:
          items:
            type: string
          type: array
          title: Warnings
        acp_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Acp Id
        acp_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Acp Version
        retry_of_job_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Retry Of Job Id
        archived:
          type: boolean
          title: Archived
          default: false
        reuse_policy:
          anyOf:
          - type: string
          - type: 'null'
          title: Reuse Policy
      type: object
      required:
      - job_id
      - state
      title: JobStatusResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError